Moved missed admin components into the Admin folder
