+++ src/app/app.module.ts
@@ -8,10 +8,10 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AdminWoodburningsComponent } from './admin/admin-woodburnings/admin-woodburnings.component';
-import { CreateWoodburningComponent } from './create-woodburning/create-woodburning.component';
-import { EditWoodburningComponent } from './edit-woodburning/edit-woodburning.component';
-import { ListWoodburningsComponent } from './list-woodburnings/list-woodburnings.component';
-import { DeleteWoodburningComponent } from './delete-woodburning/delete-woodburning.component';
+import { CreateWoodburningComponent } from './admin/create-woodburning/create-woodburning.component';
+import { EditWoodburningComponent } from './admin/edit-woodburning/edit-woodburning.component';
+import { ListWoodburningsComponent } from './admin/list-woodburnings/list-woodburnings.component';
+import { DeleteWoodburningComponent } from './admin/delete-woodburning/delete-woodburning.component';
 import { PreviewWoodburningComponent } from './preview-woodburning/preview-woodburning.component';
 import { UserProfileComponent } from './user-profile/user-profile.component';
 import { WoodburningStoreService } from './woodburning-store.service';
